add test for smhi data being same as on site and update data
